Fix query caching issue
Rails docs "query caches are created at the start of an action and destroyed at the end of that action and thus persist only for the duration of the action."

Since a cached query is not run again, execute and exec_query were not called (so clear_query_cache in the adapter extension was never called, although we switched the tenant/schema).
The problem only occurred if we ran the same query twice in a single action (i.e. Post.find(1) in public -> nil in a prepend_before_action, then Kari.switch! in before_action, then Post.find(1) -> nil in the normal action, since the query was cached previously on the public schema).